cvat/requirements/production.txt
-r base.txt
psycopg2-binary==2.7.4
mod-wsgi==4.6.2
django-basicauth==0.5.1
